A new component was generated for each product (Refactoring ProductList component)
